init.qcom.usb.sh: Remove obsolete socinfo paths
The socinfo path has not existed under /sys/devices/system/soc/
for quite a few kernel versions now, so drop support for it.
While at it, move the command output redirections to /dev/null
to within the backticks (``) since otherwise it doesn't work and
errors still get printed to stderr.

Change-Id: Ifd90d5a6d2ad34eedef38227cdae249f05baf79c